fix(chatd): restore whole-micro chat cost contracts

+++ coderd/database/dump.sql
@@ -1219,7 +1219,7 @@ CREATE TABLE chat_messages (
     context_limit bigint,
     compressed boolean DEFAULT false NOT NULL,
     created_by uuid,
-    total_cost_micros numeric
+    total_cost_micros bigint
 );
 
 CREATE SEQUENCE chat_messages_id_seq
